+++ src/merge/grouped_ptirs.rs
@@ -538,7 +538,7 @@ impl GroupedPTIR {
             self.all_canonical_ptir_counts + self.no_all_canonical_ptir_counts
         )?;
 
-        // ISOM_SRC
+
 
         let mut source_txs = tx_type_members;
         source_txs.sort_by(|left, right| {