+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 ## Changes
 
-- improve gene id assignment in merge step
+- set 0 as the default value for `-u`
 - add source gene id to the `ISOM_SRC` attribute in the merged GTF (breaking: merged GTFs from earlier versions can no longer be read by `tools revert`)